feat: initial lemmas.
The issue I have right now is that `well_formed` fixes the
order used in the tree, which is an issue if it has to interact
with `symm` since `symm` inverts the order used in `symm` (from
`<` we go to `>`). Therefore, this may require a bit more of
refactoring than originally thought.